+++ kotlinx-coroutines-core/common/test/LimitedParallelismSharedTest.kt
@@ -1,6 +1,8 @@
 package kotlinx.coroutines
 
 import kotlinx.coroutines.testing.*
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.EmptyCoroutineContext
 import kotlin.test.*
 
 class LimitedParallelismSharedTest : TestBase() {
@@ -27,5 +29,31 @@ class LimitedParallelismSharedTest : TestBase() {
         assertFailsWith<IllegalArgumentException> { Dispatchers.Default.limitedParallelism(-1) }
         assertFailsWith<IllegalArgumentException> { Dispatchers.Default.limitedParallelism(Int.MIN_VALUE) }
         Dispatchers.Default.limitedParallelism(Int.MAX_VALUE)
+    }
+
+    /**
+     * Checks that even if the dispatcher sporadically fails, the limited dispatcher will still allow reaching the
+     * target parallelism level.
+     */
+    @Test
+    fun testLimitedParallelismOfOccasionallyFailingDispatcher() {
+        val limit = 5
+        var doFail = false
+        val workerQueue = mutableListOf<Runnable>()
+        val limited = object: CoroutineDispatcher() {
+            override fun dispatch(context: CoroutineContext, block: Runnable) {
+                if (doFail) throw TestException()
+                workerQueue.add(block)
+            }
+        }.limitedParallelism(limit)
+        repeat(6 * limit) {
+            try {
+                limited.dispatch(EmptyCoroutineContext, Runnable { /* do nothing */ })
+            } catch (_: DispatchException) {
+                // ignore
+            }
+            doFail = !doFail
+        }
+        assertEquals(limit, workerQueue.size)
     }
 }